# packages/contracts/openapi/careos-api.v1.json
{
  "components": {
    "schemas": {
...
      "CreateOrganisationRequest": {
        "properties": {
          "admin_email": {
            "format": "email",
            "title": "Admin Email",
            "type": "string"
          },
          "admin_full_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Admin Full Name",
            "type": "string"
          },
          "admin_password": {
            "maxLength": 256,
            "minLength": 12,
            "title": "Admin Password",
            "type": "string"
          },
          "country_code": {
            "default": "GB",
            "pattern": "^[A-Z]{2}$",
            "title": "Country Code",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "maxLength": 80,
            "minLength": 2,
            "pattern": "^[a-z0-9][a-z0-9\\-]*$",
            "title": "Slug",
            "type": "string"
          },
          "timezone": {
            "default": "Europe/London",
            "maxLength": 64,
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug",
          "admin_email",
          "admin_full_name",
          "admin_password"
        ],
        "title": "CreateOrganisationRequest",
        "type": "object"
      },
      "CreateUserRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "email",
          "full_name",
          "role",
          "password"
        ],
        "title": "CreateUserRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrganisationStatus": {
        "enum": [
          "ACTIVE",
          "SUSPENDED"
        ],
        "title": "OrganisationStatus",
        "type": "string"
      },
      "OrganisationView": {
        "properties": {
          "country_code": {
            "title": "Country Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OrganisationStatus"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "status",
          "country_code",
          "timezone"
        ],
        "title": "OrganisationView",
        "type": "object"
      },
...
      "Role": {
        "enum": [
          "PLATFORM_ADMIN",
          "ORGANISATION_ADMIN",
          "CARE_MANAGER",
          "OPERATOR",
          "CAREGIVER",
          "TRUSTED_CONTACT"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "ServiceUserProfile": {
        "properties": {
          "address_line1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line1"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "contacts": {
            "items": {
              "$ref": "#/components/schemas/TrustedContactView"
            },
            "title": "Contacts",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "devices": {
            "items": {
              "$ref": "#/components/schemas/DeviceView"
            },
            "title": "Devices",
            "type": "array"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "escalation_policy_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Policy Id"
          },
          "external_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Reference"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "phone_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number"
          },
          "postcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postcode"
          },
          "preferred_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Name"
          },
          "recent_incidents": {
            "items": {
              "$ref": "#/components/schemas/IncidentSummary"
            },
            "title": "Recent Incidents",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/ServiceUserStatus"
          }
        },
        "required": [
          "id",
          "display_name",
          "first_name",
          "last_name",
          "preferred_name",
          "phone_number",
          "address_line1",
          "city",
          "postcode",
          "external_reference",
          "status",
          "escalation_policy_id",
          "created_at",
          "contacts",
          "devices",
          "recent_incidents"
        ],
        "title": "ServiceUserProfile",
        "type": "object"
      },
...
      "ServiceUserStatus": {
        "enum": [
          "ACTIVE",
          "INACTIVE"
        ],
        "title": "ServiceUserStatus",
        "type": "string"
      },
      "ServiceUserSummary": {
        "properties": {
          "active_incident_count": {
            "title": "Active Incident Count",
            "type": "integer"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ServiceUserStatus"
          }
        },
        "required": [
          "id",
          "display_name",
          "first_name",
          "last_name",
          "city",
          "status",
          "active_incident_count"
        ],
        "title": "ServiceUserSummary",
        "type": "object"
      },
      "ServiceUserWrite": {
        "properties": {
          "address_line1": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line1"
          },
          "city": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "escalation_policy_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Policy Id"
          },
          "external_reference": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Reference"
          },
          "first_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name",
            "type": "string"
          },
          "home_latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Latitude"
          },
          "home_longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Longitude"
          },
          "last_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name",
            "type": "string"
          },
          "phone_number": {
            "anyOf": [
              {
                "pattern": "^\\+?[0-9 ()\\-]{6,24}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number"
          },
          "postcode": {
            "anyOf": [
              {
                "maxLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postcode"
          },
          "preferred_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Name"
          },
          "status": {
            "$ref": "#/components/schemas/ServiceUserStatus",
            "default": "ACTIVE"
          }
        },
        "required": [
          "first_name",
          "last_name"
        ],
        "title": "ServiceUserWrite",
        "type": "object"
      },
...
      "TrustedContactView": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "phone_number": {
            "title": "Phone Number",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "relationship": {
            "title": "Relationship",
            "type": "string"
          }
        },
        "required": [
          "id",
          "full_name",
          "relationship",
          "phone_number",
          "email",
          "priority",
          "is_active"
        ],
        "title": "TrustedContactView",
        "type": "object"
      },
      "TrustedContactWrite": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "phone_number": {
            "pattern": "^\\+?[0-9 ()\\-]{6,24}$",
            "title": "Phone Number",
            "type": "string"
          },
          "priority": {
            "maximum": 9.0,
            "minimum": 1.0,
            "title": "Priority",
            "type": "integer"
          },
          "relationship": {
            "maxLength": 60,
            "minLength": 1,
            "title": "Relationship",
            "type": "string"
          }
        },
        "required": [
          "full_name",
          "relationship",
          "phone_number",
          "priority"
        ],
        "title": "TrustedContactWrite",
        "type": "object"
      },
      "UserView": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "organisation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organisation Id"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "organisation_id",
          "is_active",
          "last_login_at"
        ],
        "title": "UserView",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "CSRF": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "GatewayKey": {
        "in": "header",
        "name": "X-CareOS-Gateway-Key",
        "type": "apiKey"
      },
      "Session": {
        "in": "cookie",
        "name": "careos_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "CareOS does not diagnose, predict illness or recommend treatment. AI features are assistive only; incident handling is deterministic.",
    "summary": "Telecare & safety incident orchestration platform",
    "title": "CareOS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/organisations/current": {
      "get": {
        "operationId": "current_organisation_v1_organisations_current_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationView"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Current Organisation",
        "tags": [
          "organisations"
        ]
      }
    },
    "/v1/platform/organisations": {
      "get": {
        "operationId": "list_organisations_v1_platform_organisations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrganisationView"
                  },
                  "title": "Response List Organisations V1 Platform Organisations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "List Organisations",
        "tags": [
          "organisations"
        ]
      },
      "post": {
        "operationId": "create_organisation_v1_platform_organisations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganisationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Create Organisation",
        "tags": [
          "organisations"
        ]
      }
    },
    "/v1/service-users": {
      "get": {
        "operationId": "list_service_users_v1_service_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServiceUserSummary"
                  },
                  "title": "Response List Service Users V1 Service Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "List Service Users",
        "tags": [
          "service users"
        ]
      },
      "post": {
        "operationId": "create_service_user_v1_service_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceUserWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUserSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Create Service User",
        "tags": [
          "service users"
        ]
      }
    },
    "/v1/service-users/{service_user_id}": {
      "get": {
        "operationId": "get_service_user_v1_service_users__service_user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "service_user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Service User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUserProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Get Service User",
        "tags": [
          "service users"
        ]
      },
      "put": {
        "operationId": "update_service_user_v1_service_users__service_user_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "service_user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Service User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceUserWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Update Service User",
        "tags": [
          "service users"
        ]
      }
    },
    "/v1/service-users/{service_user_id}/contacts": {
      "post": {
        "operationId": "add_contact_v1_service_users__service_user_id__contacts_post",
        "parameters": [
          {
            "in": "path",
            "name": "service_user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Service User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrustedContactWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustedContactView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Add Contact",
        "tags": [
          "service users"
        ]
      }
    },
    "/v1/service-users/{service_user_id}/contacts/{contact_id}": {
      "put": {
        "operationId": "update_contact_v1_service_users__service_user_id__contacts__contact_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "service_user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Service User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "contact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Contact Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrustedContactWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustedContactView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Update Contact",
        "tags": [
          "service users"
        ]
      }
    },
...
    "/v1/users": {
      "get": {
        "operationId": "list_users_v1_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserView"
                  },
                  "title": "Response List Users V1 Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "create_user_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Create User",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}